services/example-workers/requirements.txt
aio-pika==9.4.1
croniter==2.0.3
fastapi==0.110.0
jsons==1.6.3
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
prometheus-client==0.20.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-json-logger==2.0.7
sqlalchemy==2.0.29
starlette==0.36.3
starlette-exporter==0.21.0
urllib3==2.2.1
uvicorn==0.29.0
yarl==1.9.4
